refactor(Monster) : MonsterHp 객체를 생성하여 hp 관련된 일 위임
